# specs/001-volunteer-care-report/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: StrayHub CRM Care Report API
  version: 0.1.0
  description: >-
    志工日常照護回報與動物近期歷程的 CRM 邊界契約。一般收容所操作由
    FastAPI 重新驗證 Session、Membership、角色與 Active Shelter Context；
    PLATFORM_ADMIN 使用平台級 PLATFORM Scope，且所有跨收容所操作都留下 Audit Record。
...
servers:
  - url: http://localhost:8000
    description: 本機開發
...
security:
  - bearerAuth: []
paths:
...
  /v1/care-report-drafts:
    post:
      tags: [Drafts]
      operationId: createCareReportDraft
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DraftCreateRequest" }
      responses:
        "201":
          description: Draft 已建立
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Draft" }
    get:
      tags: [Drafts]
      operationId: listCareReportDrafts
      responses:
        "200":
          description: 本人可見 Draft
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DraftListResponse" }
  /v1/care-report-drafts/{draftId}:
    parameters:
      - $ref: "#/components/parameters/DraftId"
    get:
      tags: [Drafts]
      operationId: getCareReportDraft
      responses:
        "200":
          description: Draft
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Draft" }
    patch:
      tags: [Drafts]
      operationId: updateCareReportDraft
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DraftUpdateRequest" }
      responses:
        "200":
          description: Draft 已更新
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Draft" }
    delete:
      tags: [Drafts]
      operationId: deleteCareReportDraft
      responses:
        "204": { description: Draft 已刪除或進入清理流程 }
...
  /v1/care-reports:
    post:
      tags: [Care Reports]
      operationId: createCareReport
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CareReportCreateRequest" }
      responses:
        "201":
          description: 原始 Care Report 已保存
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CareReport" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/care-reports/{reportId}:
    parameters:
      - $ref: "#/components/parameters/ReportId"
    get:
      tags: [Care Reports]
      operationId: getCareReport
      responses:
        "200":
          description: Care Report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CareReport" }
    patch:
      tags: [Care Reports]
      operationId: correctCareReport
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CareReportCorrectionRequest" }
      responses:
        "200":
          description: Care Report 已修正或封存
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CareReport" }
  /v1/care-reports/{reportId}/archive:
    post:
      tags: [Care Reports]
      operationId: archiveCareReport
      parameters:
        - $ref: "#/components/parameters/ReportId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArchiveRequest" }
      responses:
        "204": { description: Care Report 已封存並留下 Audit Record }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 簽署的短效 Access Token；每次受保護 Request 仍必須通過 Server-side Session 與 Organization Scope 驗證。
  parameters:
...
    DraftId:
      {
        name: draftId,
        in: path,
        required: true,
        schema: { type: string, format: uuid },
      }
...
    ReportId:
      {
        name: reportId,
        in: path,
        required: true,
        schema: { type: string, format: uuid },
      }
...
    IdempotencyKey:
      {
        name: Idempotency-Key,
        in: header,
        required: true,
        schema: { type: string, minLength: 1 },
      }
...
  responses:
...
    Conflict:
      description: Idempotency、stale version、operation payload 或資料狀態衝突
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ErrorResponse" } },
        }
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
        operation_id: { type: [string, "null"], format: uuid }
        details: { type: object, additionalProperties: true }
...
    DraftCreateRequest:
      type: object
      required: [animal_id, confirmation_token]
      properties:
        animal_id: { type: string, format: uuid }
        confirmation_token: { type: string, minLength: 1 }
    CareCompletionCode:
      type: string
      enum:
        [
          care_completion.completed,
          care_completion.partially_completed,
          care_completion.not_provided,
          care_completion.not_observed,
          care_completion.uncertain,
        ]
      description: 照護完成狀態；未觀察與無法判斷不得解讀為已完成
    WalkCompletionCode:
      type: string
      enum:
        [
          walk_completion.completed,
          walk_completion.partially_completed,
          walk_completion.not_done,
          walk_completion.not_observed,
          walk_completion.uncertain,
        ]
      description: 散步完成狀態；與散步反應的 walk.* 選項分開
    ObservationAnswerCode:
      type: string
      minLength: 1
      description: 由 CRM 有效 Observation Vocabulary 提供的穩定 Code；顯示名稱不得作為正式值
    DraftAnswers:
      type: object
      additionalProperties: true
      properties:
        care_completion: { $ref: "#/components/schemas/CareCompletionCode" }
        walk_completion: { $ref: "#/components/schemas/WalkCompletionCode" }
        feeding: { $ref: "#/components/schemas/ObservationAnswerCode" }
        water: { $ref: "#/components/schemas/ObservationAnswerCode" }
        activity: { $ref: "#/components/schemas/ObservationAnswerCode" }
        urination: { $ref: "#/components/schemas/ObservationAnswerCode" }
        defecation: { $ref: "#/components/schemas/ObservationAnswerCode" }
        resource_guarding:
          { $ref: "#/components/schemas/ObservationAnswerCode" }
        human_interaction:
          { $ref: "#/components/schemas/ObservationAnswerCode" }
        animal_interaction:
          { $ref: "#/components/schemas/ObservationAnswerCode" }
        emotion: { $ref: "#/components/schemas/ObservationAnswerCode" }
        walk_reaction: { $ref: "#/components/schemas/ObservationAnswerCode" }
        appearance_special_status:
          { $ref: "#/components/schemas/ObservationAnswerCode" }
      description: 可逐題保存的標準回報答案；草稿階段允許不完整，正式送出時必須符合 CareReportAnswers
    CareReportAnswers:
      allOf:
        - $ref: "#/components/schemas/DraftAnswers"
        - type: object
          required:
            [
              care_completion,
              walk_completion,
              feeding,
              water,
              activity,
              urination,
              defecation,
              resource_guarding,
              human_interaction,
              animal_interaction,
              emotion,
              walk_reaction,
              appearance_special_status,
            ]
      description: 正式 Care Report 必須完整包含的標準結構化答案；照片與心得不在此必要集合
    DraftUpdateRequest:
      type: object
      properties:
        answers:
          $ref: "#/components/schemas/DraftAnswers"
          description: 可只包含目前新增或修改的答案；後端依 Draft 目前狀態驗證，不接受跳步或略過必要題目
        note: { type: string }
        media_ids: { type: array, items: { type: string, format: uuid } }
    Draft:
      type: object
      required:
        [
          id,
          opaque_token,
          organization_id,
          animal_id,
          volunteer_user_id,
          membership_id,
          current_step,
          answers,
          status,
          expires_at,
        ]
      properties:
        id: { type: string, format: uuid }
        opaque_token: { type: [string, "null"], minLength: 1 }
        organization_id: { type: string, format: uuid }
        animal_id: { type: string, format: uuid }
        volunteer_user_id: { type: string, format: uuid }
        membership_id: { type: string, format: uuid }
        current_step:
          {
            type: string,
            enum:
              [
                selecting_animal,
                confirming_animal,
                answering_completion,
                answering_feeding,
                answering_water,
                answering_activity,
                answering_elimination,
                answering_behavior,
                answering_special_status,
                awaiting_media,
                awaiting_note,
                reviewing,
                submitting,
                submitted,
                cancelled,
                expired,
              ],
          }
        answers:
          $ref: "#/components/schemas/DraftAnswers"
        media_ids: { type: array, items: { type: string, format: uuid } }
        last_interaction_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        status: { type: string, enum: [active, submitted, cancelled, expired] }
    DraftListResponse:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Draft" } }
...
    CareReportCreateRequest:
      type: object
      required: [draft_id, observations]
      properties:
        draft_id: { type: string, format: uuid }
        observations:
          $ref: "#/components/schemas/CareReportAnswers"
          description: 標準回報完整結構化答案；照片與心得不屬於必要欄位
        note: { type: string }
        media_ids: { type: array, items: { type: string, format: uuid } }
    CareReportCorrectionRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
        observations: { type: object, additionalProperties: true }
        note: { type: string }
        animal_id: { type: string, format: uuid }
    ArchiveRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
    CareReport:
      type: object
      required:
        [id, organization_id, animal_id, status, observations, created_at]
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        animal_id: { type: string, format: uuid }
        status: { type: string, enum: [saved, amended, archived] }
        observations:
          $ref: "#/components/schemas/CareReportAnswers"
        observation_snapshots:
          type: [object, "null"]
          additionalProperties:
            type: object
            additionalProperties: { type: string }
        created_at: { type: string, format: date-time }
...